Rename directory to localproject to be consistent with the package name
