+++ ecs-cli/modules/cli/local/localproject/project_test.go
